very wip: slim down and refactor browser package
